♻️ (core): Rename usecase folder by use-case
https://ledgerhq.atlassian.net/wiki/spaces/WXP/pages/4502650957/SDK+TS+Guidelines#Folder-naming
