+++ packages/core/src/internal/config/use-case/GetSdkVersionUseCase.test.ts
